(refactor) O3-5739: Use textarea-based HTML entity decoding, move to utils
Replace the DOMParser-based decodeHtmlEntities with a detached <textarea>,
which decodes character references without parsing tags or hoisting head
content. Move the helper out of sticky-note.resource.ts into a feature-scoped
utils.ts, and decode the existing note value once in the modal instead of on
every keystroke.
