+++ packages/esm-patient-notes-app/src/sticky-notes/sticky-note-panel.component.tsx
@@ -2,7 +2,8 @@ import React, { useCallback, useMemo } from 'react';
 import { Button, SkeletonText } from '@carbon/react';
 import { useTranslation } from 'react-i18next';
 import { formatDate, showModal } from '@openmrs/esm-framework';
-import { decodeHtmlEntities, type StickyNoteObs } from './sticky-note.resource';
+import { type StickyNoteObs } from './sticky-note.resource';
+import { decodeHtmlEntities } from './utils';
 import DeleteStickyNote from './delete-sticky-note-button.component';
 import EditStickyNote from './edit-sticky-note-button.component';
 import styles from './sticky-note-panel.scss';
